+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changes
 Unreleased
 ----------
 
-- Sync official test vectors. `#64 <https://github.com/dajiaji/pyseto/pull/64 >`__
+- Sync official test vectors. `#64 <https://github.com/dajiaji/pyseto/pull/64>`__
 
 Version 1.1.0
 -------------